+++ IntegraHawkRemote/giro.h
@@ -5,17 +5,18 @@
 #define degconvert 57.2957786 //there are like 57 degrees in a radian.
 
 #if defined(ARDUINO) && ARDUINO >= 100
-	#include "Arduino.h"
+#include "Arduino.h"
+#include "Angle.h"
 #else
-	#include "WProgram.h"
+#include "WProgram.h"
 #endif
 
 class Giro
 {
- protected:
+protected:
 
 
- public:
+public:
 	Giro();
 	void init();
 	Angle GetAngles();
@@ -23,7 +24,7 @@ private:
 	int MPU_addr = 0x68;
 	double AcX, AcY, AcZ, Tmp, GyX, GyY, GyZ;
 	uint32_t timer;
-	int compAngleX, compAngleY;
+	double compAngleX, compAngleY;
 };
 
 /*extern GiroClass Giro;*/
